backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
alembic
psycopg2-binary

